+++ tests/atcoder.rs
@@ -179,10 +179,10 @@ GET https://atcoder.jp/contests/practice/submissions/me?page=1 ... 200 OK"#,
 }
 
 #[test]
-fn it_fails_to_submit_if_the_lang_id_is_invalid() -> Fallible<()> {
+fn it_fails_to_submit_if_the_lang_name_is_invalid() -> Fallible<()> {
     let _ = env_logger::try_init();
     service::test_in_tempdir(
-        "it_fails_to_submit_if_the_lang_id_is_invalid",
+        "it_fails_to_submit_if_the_lang_name_is_invalid",
         &credentials_as_input()?,
         |mut app| -> Fallible<()> {
             static CODE: &[u8] = b"#";
@@ -198,7 +198,7 @@ fn it_fails_to_submit_if_the_lang_id_is_invalid() -> Fallible<()> {
                     no_check_duplication: false,
                     service: Some(ServiceKind::Atcoder),
                     contest: Some("practice".to_owned()),
-                    language: Some("python3-with-invalid-lang-ids".to_owned()),
+                    language: Some("python3-with-invalid-lang-names".to_owned()),
                     jobs: None,
                     color_choice: AnsiColorChoice::Never,
                     problem: "a".to_owned(),
@@ -206,9 +206,9 @@ fn it_fails_to_submit_if_the_lang_id_is_invalid() -> Fallible<()> {
                 .unwrap_err();
             if_chain! {
                 if let snowchains::Error::Service(ServiceError::Context(ctx)) = &err;
-                if let ServiceErrorKind::NoSuchLangId(lang_id) = ctx.get_context();
+                if let ServiceErrorKind::NoSuchLang(lang_name) = ctx.get_context();
                 then {
-                    assert_eq!(lang_id, "invalid");
+                    assert_eq!(lang_name, "invalid");
                     Ok(())
                 } else {
                     Err(err.into())
@@ -256,10 +256,10 @@ if __name__ == '__main__':
 }
 
 #[test]
-fn it_lists_language_ids() -> Fallible<()> {
+fn it_lists_languages() -> Fallible<()> {
     let _ = env_logger::try_init();
     service::test_in_tempdir(
-        "it_lists_language_ids",
+        "it_lists_languages",
         &credentials_as_input()?,
         |mut app| -> Fallible<()> {
             app.run(Opt::ListLangs {