+++ tests/yukicoder.rs
@@ -44,10 +44,10 @@ fn it_downloads_testcases() -> Fallible<()> {
 }
 
 #[test]
-fn it_fails_to_submit_if_the_lang_id_is_invalid() -> Fallible<()> {
+fn it_fails_to_submit_if_the_lang_name_is_invalid() -> Fallible<()> {
     let _ = env_logger::try_init();
     service::test_in_tempdir(
-        "it_fails_to_submit_if_the_lang_id_is_invalid",
+        "it_fails_to_submit_if_the_lang_name_is_invalid",
         &format!("{}\n", service::env_var("YUKICODER_REVEL_SESSION")?),
         |mut app| -> Fallible<()> {
             static CODE: &[u8] = b"#";
@@ -63,7 +63,7 @@ fn it_fails_to_submit_if_the_lang_id_is_invalid() -> Fallible<()> {
                     no_check_duplication: false,
                     service: Some(ServiceKind::Yukicoder),
                     contest: Some("no".to_owned()),
-                    language: Some("python3-with-invalid-lang-ids".to_owned()),
+                    language: Some("python3-with-invalid-lang-names".to_owned()),
                     jobs: None,
                     color_choice: AnsiColorChoice::Never,
                     problem: "9000".to_owned(),
@@ -71,9 +71,9 @@ fn it_fails_to_submit_if_the_lang_id_is_invalid() -> Fallible<()> {
                 .unwrap_err();
             if_chain! {
                 if let snowchains::Error::Service(ServiceError::Context(ctx)) = &err;
-                if let ServiceErrorKind::NoSuchLangId(lang_id) = ctx.get_context();
+                if let ServiceErrorKind::NoSuchLang(lang_name) = ctx.get_context();
                 then {
-                    assert_eq!(lang_id, "invalid");
+                    assert_eq!(lang_name, "invalid");
                     Ok(())
                 } else {
                     Err(err.into())
@@ -114,10 +114,10 @@ fn it_submits_to_no_9000() -> Fallible<()> {
 }
 
 #[test]
-fn it_list_language_ids() -> Fallible<()> {
+fn it_list_languages() -> Fallible<()> {
     let _ = env_logger::try_init();
     service::test_in_tempdir(
-        "it_list_language_ids",
+        "it_list_languages",
         &format!("{}\n", service::env_var("YUKICODER_REVEL_SESSION")?),
         |mut app| -> Fallible<()> {
             static MASK_USERNAME: Lazy<Regex> = sync_lazy!(Regex::new("Username: .*").unwrap());